fixed authentication error while signin/signup
AuthController threw exception when signin/signup request was made because the ROLE entity was empty. To fix it added RoleController RoleService and RoleInitializer which needs to initialize the roles before signin/signup requests are made using /auth/test/roles. It needs to be done just once when setting up the application/DB.